Deprecate some preferences
Add a way to mark a preference deprecated without removing or otherwise
disabling it. Deprecated preference will be highlighted as such in the
command line options list, in the user manual and in the GUI profile
editor. When the user includes a deprecated preference on the command
line or in a profile then a warning is displayed, urging the user to
stop using this preference.

Deprecate the following preferences:

  - backups
  - stream
  - halfduplex