Mark some preferences as CLI only
Make it possible to mark a preference as CLI only. CLI only preferences
will not be sent to server and they are not allowed in profiles. If a
client sends a value for CLI only preference to server then it will be
silently ignored.

Mark the following preferences as CLI only:

  - version
  - doc
  - prefsdocs
  - server
  - socket
  - listen (and alias host)
  - ui
  - selftest
  - testserver
  - i
  - dumparchives