Group preferences into categories
The number of preferences is impressive... and overwhelming to users.
They are grouped into basic and advanced preferences but other than
that, there is no help to users trying to navigate the alphabetic lists
of preferences.

This patch adds a category to each preference and uses that category to
display the list of preferences in sections (and for some internal
filterting).