Fix a race condition in enableFlowControl
[enableFlowControl] calling [flushBuffer] is not safe because it calls
[sendOutput], which is not reentrant. Normally this is not a problem
because there is nothing to flush, but in some rare occasions (seems to
only happen on Windows) this will cause sending the output buffer twice.

Since the output queue is being processed already, replace the explicit
call to [flushBuffer] with a loop waiting for the output buffer to be
drained. This is done only on the client because the server output
queues are popped after flow control has been enabled.