provide packages in dune-project
Without a global dune-project file it is not possible to build just
specific libraries or executables. Instead dune will build all
libraries and executables which are listed in the individual 'dune'
files.

For example, to build just the textversion and the fsmonitor:
  dune build @install -p unison,unison-fsmonitor
  dune install unison unison-fsmonitor --destdir=/some/path/offset

Furthermore a dune-project files allows to generate opam files, which
could then be submitted to opam-repository. However, this feature is
currently disabled with (generate_opam_files false) to avoid dirtying
the existing unison.opam file.

The listed dependencies are taken verbatim from the current
unison.opam file. They have no meaning during a build with dune, but
the values listed here will be written into generated opam files.

At this point is unfortunately not yet supported to have no generated
files, such as *.opam, in the SCM.

Signed-off-by: Olaf Hering <olaf@aepfle.de>