Lwt: Fix async connect and accept on Windows
The current design keeps an array for all ongoing (initiated but not
completed) accept and connect calls. The elements of that array would
then be checked in the main run loop to see which of these calls have
completed.

However, the array-based design can't work because it expects a LIFO
usage where the most recently started connect/accept call must complete
first. If this doesn't happen then the array is immediately broken
because the count of ongoing calls is increased/decreased with every
new/completed call but the array elements are never shuffled around. (On
top of that, the array was managed in the C code and array indexes in
OCaml code.) This creates an array with "holes" in it and with some
elements stored in positions beyond the count of active elements.
Windows API is of course not happy with such input.

While this design has been used since the introduction of async Lwt on
Windows, it has really only ever worked when there is one such ongoing
call at any time. That's because due to the usage pattern of accept and
collect calls (never in parallel), the flaw was never exposed. In some
cases it may have worked with two parallel calls but then as a mere
lucky coincidence.

The flaw was recently exposed when another bug was fixed (make IPv6 work
on Windows) which caused two listening sockets to be created by the
server by default and thus two parallel accept calls.

As a fix, replace the array with a very simple solution which tracks
ongoing calls in a list. List elements are added and removed as needed,
the list remains contiguous. The list is then converted into a fresh
array for every Windows API call.
Won't this become slow? No. In practice, the number of simultaneous
ongoing calls is almost always going to be only 1 or 2 at most. Most of
the time, it will be zero (on the client and on server which is not in
socket mode).