Fix a bug with empty paths in stasher
An empty path within a root refers to the root itself. The stasher
creates and searches backup names derived from the path name. This
derivation was broken for empty paths (that is, for backing up and
merging roots, not files/dirs within roots). It is fixed by using the
root's actual name for deriving backup names when the path is empty.

The "local" backup location is chosen to be the root itself, which the
stasher expects to be a directory. Local backup location was broken for
empty paths (again, this happens when backing up/merging the root
itself). It is fixed by using to root's parent dir as the backup
location for the root when the path is empty.