Detect integer overflow for RPC message length
RPC message length is encoded in 32 bits on the wire. While certainly
extremely rare, it is not unheard of that a single RPC message may grow
beyond that limit. There is a documented case of a user trying to sync
over 120 000 000 updates in one go. The list of updates (sent in a
single RPC message) with this number of items can easily pass 4 GB
serialized.

For now, simply detect this issue and recommend the user to limit the
size of the sync. It is not practical to change the encoding of message
length given the extremes required to overflow it.