Sync the same time for merged files
When files are merged, currently their original mtimes are left
untouched on disk. However, the mtimes in the archive are updated to the
moment of the merge. This makes the merged files (now with the same
contents) conflict at the next sync because the times are now different
from the time stored in the archive and different to each other.

Fix this by updating mtimes on disk to match the mtimes in the archive.