Makefile: add manpage to "all" target
This is a point fix, avoiding the (needed) grand restructuring, but it
enables packaging systems that want to build by calling a single
target to build the unison binary and the man page.  Basically, I'm
taking the view that no one should want to avoid the man page, at
least by default.