src/Makefile.OCaml: Avoid nonportable usage in dot checking
Previously, there was nonportable usage (type -t, bash) in checking
for ocamldot, and ocamldot was run if present on "make depend".

Separate preparation of the graph as "dependgraph".

Remove the checks, and simply attempt to run ocamldot and dot, and
ignore failures.  If the tools are present, DEPENDENCIES.ps will be
created, and if they aren't, there will be error messages.