Add tests for additional RPC coverage
The self-tests currently do not cover the entire range of RPC API
functions. Add tests for "fingerprintSubfile", "updateProps", "mkdir",
"setupTargetPaths" and "replaceArchive" RPC functions. Together with the
pre-existing tests, all primary RPC functions are now tested at least
once.

Note that after adding these tests, the RPC coverage is still not
complete as some functions are too difficult to test with the current
testing framework or otherwise not critical. The tests do not cover all
possible ranges of parameter values for the functions.