Add workaround for [readlink] weirdness in Windows
Under some circumstances (unclear which exactly; always?) [readlink]
apparently does not produce clean results in Windows. Links with an
absolute path to target have been re-formatted. It looks like absolute
paths are presented in the NT namespace, under \DosDevices, which has a
shorthand of \?? (present as prefix \??\). It is not fully clear if
paths under \DosDevices in the NT namespace are the same as paths in
either of Win32 file namespace or Win32 device namespace, but for our
purposes here it is sufficient to consider that it so (with the Win32
file namespace being the only one of interest). This is also confirmed
by MSDN at least for the Win32 device namespace
https://docs.microsoft.com/en-us/windows-hardware/drivers/kernel/introduction-to-ms-dos-device-names
https://docs.microsoft.com/en-us/windows-hardware/drivers/kernel/object-directories
https://docs.microsoft.com/en-us/windows-hardware/drivers/kernel/object-names

Paths in NT namespace are not useful for us. To work around that and
produce a working path, even if it may not result in 100% the same value
that was set by [symlink] (which are mostly expected to be in DOS or UNC
format), simply change the prefix \??\ to the Win32 file namespace
prefix \\?\.