Improve future OCaml compatibility
Names of some public symbols have changed in OCaml's Unix library. The
change will be first released with OCaml 5.0.  Compatibility macros are
provided for some changed names but not for all.

Use the new names in the C code and add local macros for compatibility
with older compilers.

This is nice-to-have on non-Windows platforms (OCaml's compatibility
macros make the build work either way) but must-have on Windows (build
would break otherwise).

Also fix an error from an earlier commit, a misspelled OCAML_VERSION.