System_impl: Remove [Fs] and [System] separation
It seems that the original intention with separating [System_impl.Fs]
and [System_impl.System] was to use the former on paths within replicas
and the latter for filesystem and system access outside the replicas. On
POSIX systems there is no difference between those two sub-modules. On
Windows, there used to be two differences:

 - first, the Win32 API to use (ANSI or Unicode) was user-configurable
   for the former sub-module, while for the latter it was always the
   Win32 Unicode API;
 - second, the former sub-module was doing some pre-processing on input
   paths while the latter passed through input paths unchanged.

The first of these differences had become outdated and was removed by
commit 336837087f785302db4473ca587e52fd812eca6d.

The commit at hand removes the second difference and thus also the
separation between [System_impl.Fs] and [System_impl.System].

The path pre-processing, while enabling full power of filesystem
functions in Windows Unicode API, turns off any path processing by
Windows API, which may or may not be desired. (See more about this at
https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file)

The reason for the complex functorized module hierarchy is to in one
case do the pre-processing automatically and transparently (this is done
for the [System_impl.Fs] sub-module), in the other case switch off this
pre-processing in order to leverage the path processing by Windows API
(for example, accepting / instead of \ as path separator, and expanding
".."). This is done for the [System_impl.System] sub-module.  (It is
actually unclear if this additional processing by Windows is really
needed anywhere in Unison code. Even if it is, it could probably be
avoided by careful coding.)

Since the [System_impl.Fs] module is used almost exclusively by the
toplevel [Fs] module, just request path pre-processing there explicitly
and simplify the entire hierarchy of [System] modules respectively.

The [Fspath] module also makes use of [System_impl.Fs] but it uses a
very limited subset of functions from that module, most of which do not
use the path pre-processing (with comments at [System_win.chdir]
explicitly stating it).