Make [System.fspath] non-abstract
[System.fspath] does not need to be an abstract type. It can be argued
whether there were good reasons for it when the module was introduced.
Perhaps it served to somewhat manage the separation of in-replica and
outside-replica usage, or manage the path pre-processing in Windows, or
just in case for a yet-unknown-future-use. While the in-replica vs
outside-replica separation sounds plausible, it only works because of
discipline in the calling code. The type being abstract or not does not
change this. As for other reasons, these are mostly gone after recent
simplifications.

Keep [System.fspath] as a separate type only for the [Fs] module. Make
it non-abstract for all other direct uses of [System] and replace all
occurrences of [System.fspath] type in other modules' signatures with
the string type.