Move charset conversion for Win32 Unicode API to C
Windows Unicode API uses 16-bit chars, so-called wide chars (UTF-16-ish
encoding). Currently the conversion to and from 8-bit chars (expected to
be UTF-8, but not necessarily) is done in OCaml code. This is not only a
burden, it is a major risk because the OCaml data type is still string
and any missed conversion will at least result in garbage data and will
potentially cause a segfault.

Move the charset conversion for Windows API into C stubs, so that all
strings in OCaml (and other C stubs) use 8-bit chars and are UTF-8
(almost true; technically they could even be in some proprietary Windows
8-bit encoding...).  Windows charset conversions are carried out only
right next to Win32 API invocations where they are needed.