Reset some state after remote connection is closed
There is some global state which is not cleaned up properly when the
remote connection is closed unexpectedly. This is usually not a problem
because in most cases both the client and server quit when the
connection breaks. Only the socket server keeps running and accepting
new clients after a client has dropped. It is also possible that in
future a client could survive the server disappearing without quitting
the client process.

This is not a problem when the connection is closed cleanly. The global
state would then have been cleared as part of the normal processing.

The fix is to reset some of this state when the connection is closed,
yet the process keeps running. Alternatively, some of the state could be
reset every time when starting a new sync. There are several arguments
against this solution, though. First, in the vast majority of cases the
syncs already start in a clean state. Next, some state (like open file
descriptors) can be reset only afterwards. And finally, resetting state
directly after a connection close allows GC to free memory immediately.
For a uniform solution, all state is reset at connection close.

Add a possibility to register functions which are run after the
connection is closed (cleanly or unexpectedly) but the process keeps
running. This is currently done only for the server but can be used for
the client in future.

Add the following state reset functions and register them to run at
connection close:

 - Close open file descriptors (called "compressors" for input files and
   "decompressors" for output and rsync reference files) used for remote
   file transfers.

 - Clear the list of pending transfers that would otherwise block new
   attempts to transfer files with the same fingerprint.

 - Clear RPC streaming mode error state.

 - Reset the Lwt region for RPC streaming, which is essentially a queue
   of pending transfers. Not resetting this queue means that no new
   transfers can start because none of the existing transfers in the
   queue will be continued and keep blocking the queue forever.