RPC streaming: Don't ping server for client errors
Note: In the following, "client" and "server" are relative terms which
don't necessarily correspond with how the user thinks of the endpoints.
Either way, these are two Unison processes communicating with each
other.

RPC streaming mode has an "error reset" mechanism which involves a
server notifying the client of the error and the client ACKing it
with a "ping", which will also clear the error state at server side.

The client used to send this "ping" on all error conditions, even for
those that did not originate on the server. This usually works as a
no-op on server.

There are some, rather rare, error conditions which prevent the client
from sending anything more to the network while at the same time can
also block the server from ever receiving that information.

There are two independent problems at play here which both individually
cause the client and server to hang, possibly indefinitely.

The first issue. When the write syscall on the network socket is
permanently interrupted (this would typically be a SIGINT, for example)
then the RPC output queue is left in an inconsistent state which blocks
any future writes. Normally, a SIGINT would kill the process but it is
possible to request receiving it as a Sys.Break exception (which also
would kill the process if uncaught, but that's not the case).

The second issue. The current RPC protocol is designed in a way which
can't handle receiving incomplete RPC packets. The receiving end will
either block indefinitely waiting for the remainder of the packet to
arrive or will break with a protocol error if more data is received
which crosses the expected packet boundary.

While the issues are completely separate they both cause a deadlock when
the client tries to send more data to the server after experiencing an
incomplete write on the network socket. Sending the "ping" above will do
exactly that, try sending more data. (Note that sending any Transient or
Fatal exception in this situation would cause a deadlock the same way.)

Fixing the RPC output queue (first issue) or the RPC protocol (second
issue) is a complex task and mostly unnecessary. Closing the connection
will clear the state and allow the sync to proceed after re-establishing
the connection. Unison is built to be resilient over connection breaks.

To close the connection before the deadlock can happen, only the trigger
to deadlock must be fixed. In this case, add a condition on the client
to send the "ping" only when requested by the server. This way we know
that the error was not caused by an interrupted write syscall.