Close the archive file on input errors
Protect the archive file descriptor when loading the archive and close
it in case of any exceptions. Not closing the file descriptor can cause
issues in Windows if the process keeps running (a socket server, for
example) The file remains locked and even "ignorearchives" will no
longer work for the user.