Fix setting filetimes before Unix epoch in Windows
Commit 542db482cbe7495690e765d432a2f675ffa02b52 removed the stub for
Windows utimes function but not for the stat function. This introduced a
slight difference in Windows filetime to/from Unix time conversion. The
stubs for stat and utimes functions used to do the same conversion as
stat and utimes in the OCaml's Unix library but with a crucial
difference: the sequence of arithmetic operations was the reverse of
that in the Unix library. This does not matter for positive time values
but breaks with negative values (unsigned integer math). Windows
filetime epoch is 1601-01-01 UTC, which is well before the Unix epoch
(meaning negative Unix time).

Make the sequence of time conversion arithmetic operations in the stub
for Windows stat function the same as that in the Unix library. The
output of [System.stat] can now be fed correctly into [Unix.utimes] even
for times before the Unix epoch.