Store more accurate props for merged files
Currently the props of the merge result file are stored in the archive
for merged files. These props may have nothing in common with the actual
files as the merge result may be a completely new file.

This works well enough in many cases because Unison still sets the
correct props on the merged files, it just does not record them in the
archive. At next scan Unison will detect that the props have changed
(since they are different from ones in the archive) and will update the
archive accordingly.

This method breaks down when there were (potentially conflicting)
changes in props of both files or when there are further (potentially
conflicting) changes after the merge and before the next scan.

The fix is to at the time of merge update the archive with props that
are as close to reality as possible. The props with no change or equal
change are selected, making any actual props changes re-appear at next
scan. If it is not possible to detect which props should be set in the
archive then the archive is left unchanged, which is better than storing
props of a completely separate file (the merge result) and will detect
the real changes at next change.

This is not as big a change as it may look at first glance. It only
makes a tangible difference in a subset of cases mentioned above. It
also sets the scene for being able to merge file properties in future.