Leverage type system to isolate usages of Props
Basic props, meaning those available by stat(2), are used much more
widely in the code than so-called extended props. Extended props are
used for actual propagation and stored in the archive, but they can be
more expensive to retrieve and compare.

Separate usages of basic props and extended props by different types to
leverage the type system and have the compiler enforce the different
usages while keeping the data structure unchanged. The different types
are phantom types with the underlying type being the same as the
previous [Props.t]. There will be no differences in the archive on the
disk or on the wire.

This paves the way for adding new extended props, like ACLs and extended
attributes, which may be very expensive to retrieve compared to a simple
[stat]. By clearly isolating when these extended props are and are not
actually needed, this cost can be reduced to a minimum.

Make similar changes to [Fileinfo.t] and add variations of
[Fileinfo.get] to make use of the different [Props.t] variations.

Change the interface of [Os.createDir] to not require [Props.t] as
input.

Change the interfaces of [Os.fingerprint] and [Osx.ressFingerprint] to
not require [Fileinfo.t] (which includes [Props.t]) and [Osx.info],
respectively. Only require [Fileinfo.typ].