Add ability to set initial values in [Pred.create]
Predicates lack a way to set overridable default values like preferences
have. There is a way to set defaults but these are persistent.

Add a way to set initial values which are set as default values for the
underlying preference created for the predicate. The user preferences
will not override the default value (since this is a repeating
preference, adding values to a list) but this allows changing the
default value in future Unison releases.