Add support for xattr synchronization
Support synchronization of extended attributes as file properties
(meaning that change in extended attributes is not reported as change in
file data, unlike for resource forks).

Extended attributes are treated as simple name-value pairs with a
maximum value size of nearly 16 MB. This limit is currently seen as
plenty adequate and chosen because many popular operating systems don't
come even close to this limit and many also don't offer a reasonable API
to deal with values larger than perhaps 1 MB. There are operating
systems which do support larger values and offer a more flexible API,
but support for this is currently not implemented.

The number of attributes on a single file is not limited. The number and
total size of extended attributes on each file is not known beforehand.
This makes it impossible to store the attributes verbatim in the
archive. Instead, attribute name and fingerprint of value (for any
values larger than 32 bytes) are stored in the archive.

The total size of attributes is recorded but currently not included
in sync progress reporting.

Add new preference to request/enable xattr synchronization.

Add ignore and ignorenot predicates for extended attributes.

Add a conditional combinator [cond] to the Umarshal module to support
altering the types in backwards and forwards compatible manner.