Load xattr values on demand
Extended attribute values larger than 32 bytes are not stored in the
archive and are not available in the [Props.t] data structure. To
propagate such attributes, they have to be loaded in the source replica
on demand when copying the file or when propagating updated properties
for an unchanged file.

Since this loading necessarily has to happen on the "other" root (not
the root the update is being propagated to), this may increase the
number of network round-trips. Some care is taken to make sure that
there are no empty network requests (when there are no values to load)
and the loading is as efficient as possible. When the global path is
resolved to a local path, this resolved path is returned and reused when
updating the source archive. "updateProps" RPC function is updated
accordingly and kept backwards compatible.