Record xattrs pref and predicates in the archive
Store the "xattrs" preference and related predicates in the archive to
force a rescan of all props when the preference or predicates change.
If a rescan is not forced then fast updates checking (by mtime and
ctime) will effectively skip any new attributes that would otherwise be
included by new user preferences.