Add optionally persistent shared data for props
Until now, all the props have been individual atoms, tied to a specific
file and not knowing any context in the larger archive scope.

This commit adds the option for props to create shared data and
optionally persist it in the archive. The shared data is not only shared
between all the props in the same replica but also between the replicas.

The first use case for shared props data is a non-persistent cache for
extended attribute values. Since the attribute values must be loaded at
least once for fingerprinting and again when propagating updates, the
cache can be used to reduce the number of times attribute values must be
loaded from disk and retrieved over the network while not increasing
size of the archive (values larger than 32 bytes are not stored in the
archive).

The cache is very primitive. It stores only small values (currently less
than 1024 bytes in size) and permits only a small number of entries
(currently 200 for one replica). This is done to reduce memory usage and
unwanted network transmission. The cache does not have any means of
keeping the most preferred values (in this case, most frequently used
ones). Once the cache is full, it's full. Until the next scan.