Make search for AppleDouble files more resilient
Search for an AppleDouble file currently passes through non-fatal errors
from [Fspath.findWorkingDir]. This can in certain cases reject valid
paths.

The key to the fix is that these errors are raised on the real file path
(not on the AppleDouble file path). This means that any relevant error
processing is already in place without AppleDouble files being involved.

The errors can and should be ignored during search for AppleDouble file.
This will not silently ignore important errors but it will allow valid
paths (such as a filesystem root) to cleanly pass through AppleDouble
processing.