Make GUI-only Windows binary more discoverable
Some output (for example -version and command line parsing errors) is
produced only as text. GUI-only Windows binary will silently hide such
output. Make it more discoverable by presenting the output as a GUI
message box. If that's not possible then explain to the user that this
is a GUI-only executable.