Sort the list of extended attributes by name
The hash of the list of extended attributes (used for checking archive
integrity) was not stable because it failed to take into account list
ordering (which is otherwise irrelevant for syncing purposes).

Sort the list of extended attributes by name when first reading it from
fs to make sure that the list ordering is the same for all replicas.