Skip more empty output from remote shell terminal
Empty output from the remote shell program (that is, output which will
be presented as a prompt to the user) is currently only detected before
any processing is done on it. The processing of ANSI escapes is done
after this detection. This can result in the output becoming empty and
the user receiving an empty prompt.

Fix this by checking for empty output after the extra processing.