Improve error handling for remote shell prompts
The Lwt thread reading from the remote shell terminal keeps running
(actually, waiting...) even after the remote connection has been
established. Since no more output is (normally) expected on the
terminal, the thread is essentially blocked forever. Except for when the
"blocking" read call returns an error. We are not interested in those
errors. The connection was already established; the real error, which
will be handled in the Remote module, is the connection breaking.

Ignore IO errors when reading from remote shell terminal once the
connection has been established.