Reduce memory usage during update propagation
Currently all update actions are propagated by up front creating a
separate Lwt thread for each update. All threads are kept in a single
list and the sync is considered complete once all threads in this list
have completed their work. This approach, creating all threads up front
and keeping them in a list, can keep a lot of memory allocated because
threads and their data are only collected by GC once the entire list is
collected. For large syncs, with hundreds of thousands or millions of
updates, the memory wasted by inactive and uncollected completed threads
can reach gigabytes.

This patch includes two changes to considerably reduce memory
requirement. First, reduce the number of Lwt threads which are not
collectable by GC by not storing the threads in a list or some other
data structure. This makes detecting sync completion a bit harder. In
this patch it is solved by adding a separate thread which acts sort of
like a ref counter and wakes up when the counter of running threads
reaches zero (or when any thread fails with an exception).

Second, don't create threads that won't be started due to the limit of
'maxthreads' (essentially the worker pool size). New threads are created
on demand when workers become available in the pool.