Clean up client connection management
Code for managing client connections in the Remote module is scattered
around and duplicated in many locations. Move the code into dedicated
functions and collect most of the functions in a new sub-module.

While at it, make some smaller code changes (no change in functionality)
to prepare for future improvements.