Simplify client connection management
... and increase its correctness just a tiny bit. In previous code,
looking up connections used a less unique key than the one used when
registering opened the connections. This resulted in the previous code
potentially overwriting distinct connections. This is still not fully
fixed. This is not really a problem at the moment because several
connections are never used in parallel.