Close sockets and trigger connection cleanup reliably
The code before recent commits never closed sockets (or remote channel
fds) and remote connections on the client. On the socket server, the
connected socket was closed only once the exception of lost connection
had reached the topmost loop of the server.

Now, when detecting a lost connection at low-level IO, the socket (or
remote channel fd) that lost the connection and other fds (if any)
grouped into the same "connection" are closed immediately. After that,
connection cleanup is triggered (running registered [at_conn_close]
handlers). Finally, an exception for lost connection is raised (as
before).

This means that any code catching the lost connection exception is run
after all the cleanup has been executed. This is expected not to be a
problem, as cleanup code will typically close open resources and purge
pending tasks, so is either a duplicate of or at least not interfering
with other exception-catching code.