Add cleanup of resources after connection close
When a connection is closed in the middle of ongoing update propagation
then the exception will stop the running propagation threads (of which
there can be thousands, as these are cheap Lwt threads).

Add cleanup code to close resources and purge pending tasks, primarily
intended to be run when a connection is unexpectedly closed. The same
cleanup will also be executed when connection is closed intentionally.

This cleanup has not been needed before because any uncaught exception
during transport was considered a fatal error and resulted in quitting
the UI. Since 3a4e1588f1a19011bdf71fddf20f0aa393e6269c, the GUI can
survive such fatal errors and will benefit from the cleanup.