Improve algorithms for rate and ETA calculation
There are a few major problems causing the overall sync progress rate
and ETA calculation to be (sometimes wildly) inaccurate:

 - Activities during a sync can be highly non-uniform and don't
   contribute proportionally towards the progress. For example,
   scanning file contents for fingerprinting is currently not counted
   towards progress, even though in some cases it could be the majority
   of work done during a sync.

 - Activities that count proportionally correctly towards progress can
   still be mixed in a non-uniform way, making the progress rate (and
   thus the ETA) highly volatile. For example, many sending and receiving
   data streams over the network are active in parallel. Asymmetric
   network bandwidth causes these mixed transfers to result in varying
   overall transfer rate, as perceived by the progress tracking code.
   On top of that, some network transfers may be skipped completely
   if there is an opportunity for a local transfer, causing an
   unpredictable mix of local and bidirectional network transfers
   contributing towards the overall progress.

The current algorithms for calculating the overall sync progress rate
(not displayed in text UI) and ETA are the simplest imaginable (slightly
more sophisticated in the GUI). This causes the ETA to be practically
non-usable in many real world sync scenarios that last for more than a
few tens of seconds.

In text UI, the progress rate is calculated simply as a global average
of total copied file size over sync time; that is, the rate is the
average over the entire time the sync has been ongoing at any moment.
Consequently, the current rate becomes more and more disconnected from
reality with every second the sync is ongoing. On the positive side,
this provides for a very stable ETA. If only it wasn't useless, that
ETA...

The GUI uses a moving average for the displayed overall rate but due to
the underlying difficulty of calculating a stable rate (see above), this
value does not lend itself for ETA calculation, so the ETA is calculated
the same way as in the text UI.

This patch implements new algorithms for calculating the overall sync
progress rate and the ETA.

There are actually three progress rates calculated: 1) the "raw" current
rate, which is the actual amount of data transferred over 0.25 seconds;
2) a moving average of the previous, which can be displayed to the user
as the current sync rate; and 3) a further smoothed moving average of the
previous (a double moving average), which can be used to calculate the
ETA. Additonally, the ETA calculation algorithm is made "fuzzy" by not
producing exact-to-the-second results, and instead decreasing the time
resolution more the longer the ETA is. For example, times over 2 minutes
are displayed in 30 second steps.

The moving averages are calculated by an exponential algorithm (aka
exponential smoothing). Additional smoothing (elimination of small
fluctuations) for the second moving average is further achieved by
applying a Gaussian filter.

The purpose of these algorithms is, without attempting to do anything
about the underlying causes, to provide the user with a realistic
progress rate that is accurate over longer time periods, yet still able
to reasonably quickly react to global sync speed changes and produce a
meaningful and reasonably stable ETA.