Add per-item progress tracking in TUI
In text UI, the overall sync progress status does not consider updates
which don't copy any file data (properties updates and deletions, for
example). This can become inconvenient for the user if the number of
such non-copy updates is large or there have been changes in the target
replica (causing a rescan for deleted files, for example). Unlike the
GUI, where per-file status is displayed in the list of updates, only the
overall progress status is displayed in the text UI.

This patch provides a slight remedy to this by displaying the completed
and total number of non-conflicting updates in the progress status. The
number of updates is only for display; it does not affect the overall
progress rate and ETA calculation.